refactor(DPGO-19): move auth/jwt to service
