+++ internal/server/http/router.go
@@ -9,8 +9,8 @@ import (
 	"github.com/ITA-Dnipro/Dp-210_Go/internal/role"
 	handlePasw "github.com/ITA-Dnipro/Dp-210_Go/internal/server/http/handlers/user/password"
 	"github.com/ITA-Dnipro/Dp-210_Go/internal/server/http/middleware"
-	"github.com/ITA-Dnipro/Dp-210_Go/internal/server/http/middleware/auth"
 	handlers "github.com/ITA-Dnipro/Dp-210_Go/internal/server/http/user"
+	"github.com/ITA-Dnipro/Dp-210_Go/internal/service/auth"
 	"github.com/ITA-Dnipro/Dp-210_Go/internal/service/sender/mail"
 	usecases "github.com/ITA-Dnipro/Dp-210_Go/internal/usecases/user"
 	usecasesPasw "github.com/ITA-Dnipro/Dp-210_Go/internal/usecases/user/password"
